Moved applyTextboxControls to useCanvasControls
