+++ composables/useCanvasControls.ts
@@ -1,11 +1,10 @@
 // composables/useCanvasControls.ts
 
-import { FabricImage, Textbox, Control, controlsUtils } from 'fabric'
+import { FabricImage, Textbox, Control, controlsUtils, util } from 'fabric'
 import type { Canvas, FabricObject } from 'fabric'
 import { CircularTextbox } from '~/utils/canvasCircularTextbox'
 import { setTextboxTextRadius, toggleObjectZOrder } from '~/utils/canvasUtils'
 import { attachSvgDataUrl } from '~/composables/useCanvasImage'
-import { useCanvasText } from '~/composables/useCanvasText'
 import {
   createResizeControlRender,
   createRotateControlRender,
@@ -31,7 +30,6 @@ interface Transform {
  */
 
 export function useCanvasControls() {
-  const { applyTextboxControls } = useCanvasText()
 
   /**
    * Apply custom controls and appearance to a FabricImage.
@@ -121,6 +119,99 @@ export function useCanvasControls() {
     })
   }
 
+  /**
+   * Apply custom controls to an existing Textbox.
+   * Called both when first adding and when restoring from JSON.
+   */
+  function applyTextboxControls(textbox: Textbox): void {
+    textbox.controls = {
+      bringToFrontIcon: new Control({
+        x: -0.5,
+        y: -0.5,
+        offsetX: -12,
+        offsetY: -12,
+        sizeX: 36,
+        sizeY: 36,
+        cursorStyle: 'pointer',
+        render: createBringToFrontControlRender(getBringToFrontImage()),
+        mouseUpHandler: (eventData, transform) => {
+          const target = transform?.target
+          if (target && target.canvas) {
+            toggleObjectZOrder(target, target.canvas)
+          }
+        }
+      }),
+      scaleIcon: new Control({
+        x: 0.5,
+        y: 0.5,
+        offsetX: 12,
+        offsetY: 12,
+        sizeX: 36,
+        sizeY: 36,
+        render: createResizeControlRender(getResizeImage()),
+        cursorStyle: 'nwse-resize',
+        actionHandler: controlsUtils.scalingEqually,
+      }),
+      deleteIcon: new Control({
+        x: 0.5,
+        y: -0.5,
+        offsetX: 12,
+        offsetY: -12,
+        sizeX: 36,
+        sizeY: 36,
+        cursorStyle: 'pointer',
+        render: createTrashControlRender(getTrashCanImage()),
+        mouseUpHandler: (eventData, transform) => {
+          const target = transform?.target
+          if (target) {
+            const canvas = target.canvas
+            canvas?.remove(target)
+            canvas?.requestRenderAll()
+          }
+        }
+      }),
+      rotateIcon: new Control({
+        x: 0,
+        y: -0.5,
+        offsetY: -50,
+        sizeX: 36,
+        sizeY: 36,
+        cursorStyle: 'grab',
+        render: createRotateControlRender(getRotateImage()),
+        withConnection: true,
+        actionHandler: controlsUtils.rotationWithSnapping,
+      }),
+      resize: new Control({
+        x: -0.5,
+        y: 0.5,
+        cursorStyle: 'ew-resize',
+        offsetX: -12,
+        offsetY: 12,
+        sizeX: 36,
+        sizeY: 36,
+        render: (ctx, left, top, _styleOverride, fabricObject) => {
+          const size = 24
+          const img = getResizeImage()
+          ctx.save()
+          ctx.translate(left, top)
+          ctx.fillStyle = 'white'
+          ctx.rotate(util.degreesToRadians(fabricObject.angle || 0))
+          ctx.beginPath()
+          ctx.arc(0, 0, (3 * size) / 4, 0, Math.PI * 2)
+          ctx.fill()
+
+          if (img.complete) {
+            ctx.drawImage(img, -size / 2, -size / 2, size, size)
+          }
+
+          ctx.restore()
+        },
+        withConnection: true,
+        actionHandler: controlsUtils.changeObjectWidth,
+      }),
+    }
+  }
+
   function reapplyControlsToObject(obj: FabricObject): void {
     if (obj instanceof FabricImage) {
       applyImageControls(obj)
@@ -160,5 +251,5 @@ export function useCanvasControls() {
     }
   }
 
-  return { applyImageControls, reapplyControls, reapplyControlsToObject, makeControlsReviver }
+  return { applyImageControls, applyTextboxControls, reapplyControls, reapplyControlsToObject, makeControlsReviver }
 }
